Changed the way login / logout is handled -- now no check done for login path
