+++ test/test_helper.exs
@@ -1,4 +1,4 @@
 ExUnit.start()
 
-redirect_pages = %{"admin" => "/admin", "user" => "/users", nil => "/"}
+redirect_pages = %{"admin" => "/admin", "user" => "/users", nil => "/", "login" => "/admin/login"}
 Application.put_env(:openmaize, :redirect_pages, redirect_pages)
